Moved Paddle / PTT handler out of uhsdr_main, other minor changes
renamed/regrouped functions in audio_filter.c
