+++ mchf-eclipse/hardware/uhsdr_board.c
@@ -34,6 +34,11 @@
 #include "dac.h"
 
 #include "uhsdr_keypad.h"
+
+//required for interrupt handler
+#include "radio_management.h"
+#include "cw_gen.h"
+
 // #include "osc_interface.h"
 
 // Transceiver state public structure
@@ -255,6 +260,44 @@ void Board_InitMinimal()
     }
 
 }
+
+/**
+ * Interrupt handler for the input paddles/ptt switch
+ * @param GPIO_Pin
+ */
+void HAL_GPIO_EXTI_Callback (uint16_t GPIO_Pin)
+{
+    if (ts.paddles_active != false)
+    {
+        switch(GPIO_Pin)
+        {
+        case BUTTON_PWR:
+            break;
+        case PADDLE_DAH:
+            // Call handler
+            if (Board_PttDahLinePressed() && RadioManagement_IsTxDisabledBy(TX_DISABLE_RXMODE) == false)
+            {  // was PTT line low? Is it not a RX only mode
+                RadioManagement_Request_TxOn();     // yes - ONLY then do we activate PTT!  (e.g. prevent hardware bug from keying PTT!)
+                if(ts.dmod_mode == DEMOD_CW || is_demod_rtty() || is_demod_psk() || ts.cw_text_entry)
+                {
+                    CwGen_DahIRQ();     // Yes - go to CW state machine
+                }
+            }
+            break;
+        case PADDLE_DIT:
+            if((ts.dmod_mode == DEMOD_CW || is_demod_rtty() || is_demod_psk() || ts.cw_text_entry) && Board_DitLinePressed())
+            {
+                if (ts.cw_keyer_mode != CW_KEYER_MODE_STRAIGHT)
+                {
+                    RadioManagement_Request_TxOn();
+                }
+                CwGen_DitIRQ();
+            }
+            break;
+        }
+    }
+}
+
 
 /*
  * This initializes non-essential hardware for later use by the application
